Setting up all files, interface, abstract classes, and classes

+++ src/zoosim/IObject.java
@@ -5,10 +5,25 @@
 package zoosim;
 
 /**
- *
+ * Interface representing an object in the zoo.
+ * Defines methods for object behavior.
+ * 
  * @author Daniel Zhong
  */
 public interface IObject {
-    public void turn(int degrees);
-    public void place(int x, int y);
+    
+    /**
+     * Turns the object by a certain angle.
+     * 
+     * @param degrees the angle by which to turn the object
+     */
+    public void turn(double degrees);
+    
+    /**
+     * Places the object at a specified position.
+     * 
+     * @param x the x-coordinate of the position
+     * @param y the y-coordinate of the position
+     */
+    public void place(double x, double y);
 }
